test: keep trace guardrail modules audit-clean
