+++ src/commands/trace/tests.rs
@@ -25,42 +25,6 @@ fn aggregate_samples(durations: &[u64]) -> Vec<TraceAggregateSpanSample> {
         .collect()
 }
 
-fn trace_args_for_rig(rig_id: &str) -> TraceArgs {
-    TraceArgs {
-        comp: PositionalComponentArgs {
-            component: Some("studio".to_string()),
-            path: None,
-        },
-        component_arg: None,
-        scenario: Some("studio-app-create-site".to_string()),
-        scenario_arg: None,
-        compare_after: None,
-        rig: Some(rig_id.to_string()),
-        setting_args: SettingArgs::default(),
-        _json: HiddenJsonArgs::default(),
-        json_summary: false,
-        report: None,
-        experiment: None,
-        repeat: 1,
-        aggregate: Some("spans".to_string()),
-        schedule: TraceSchedule::Grouped,
-        focus_spans: Vec::new(),
-        spans: Vec::new(),
-        phases: Vec::new(),
-        phase_preset: None,
-        baseline_args: BaselineArgs::default(),
-        regression_threshold: extension_trace::baseline::DEFAULT_REGRESSION_THRESHOLD_PERCENT,
-        regression_min_delta_ms: extension_trace::baseline::DEFAULT_REGRESSION_MIN_DELTA_MS,
-        overlays: Vec::new(),
-        variants: Vec::new(),
-        matrix: TraceVariantMatrixMode::None,
-        output_dir: None,
-        keep_overlay: false,
-        stale: false,
-        force: false,
-    }
-}
-
 #[test]
 fn rig_component_path_and_trace_env_are_threaded() {
     let component_dir = tempfile::TempDir::new().expect("component dir");
@@ -328,107 +292,6 @@ fn rig_trace_run_uses_rig_owned_workload_extension_without_component_link() {
                 );
             }
             _ => panic!("expected run output"),
-        }
-    });
-}
-
-#[test]
-fn trace_aggregate_runs_passing_guardrails() {
-    with_isolated_home(|home| {
-        let _xdg = XdgGuard::without_xdg_data_home();
-        write_trace_extension(home);
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig(home, "studio-rig", "studio", component_dir.path());
-
-        let rig_path = home
-            .path()
-            .join(".config")
-            .join("homeboy")
-            .join("rigs")
-            .join("studio-rig.json");
-        let mut rig_json: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&rig_path).expect("read rig"))
-                .expect("parse rig");
-        rig_json["trace_guardrails"] = serde_json::json!([
-            { "label": "behavior smoke", "command": "true" }
-        ]);
-        fs::write(
-            &rig_path,
-            serde_json::to_string_pretty(&rig_json).expect("serialize rig"),
-        )
-        .expect("write rig");
-
-        let (output, exit_code) =
-            run_repeat(trace_args_for_rig("studio-rig")).expect("trace aggregate should run");
-
-        assert_eq!(exit_code, 0);
-        match output {
-            TraceCommandOutput::Aggregate(aggregate) => {
-                assert!(aggregate.passed);
-                assert_eq!(aggregate.guardrail_failure_count, 0);
-                assert_eq!(aggregate.guardrails.len(), 1);
-                assert_eq!(aggregate.guardrails[0].label, "behavior smoke");
-                assert!(aggregate.guardrails[0].passed);
-            }
-            _ => panic!("expected aggregate output"),
-        }
-    });
-}
-
-#[test]
-fn trace_aggregate_fails_guardrails_without_losing_timing_artifacts() {
-    with_isolated_home(|home| {
-        let _xdg = XdgGuard::without_xdg_data_home();
-        write_trace_extension(home);
-        let component_dir = tempfile::TempDir::new().expect("component dir");
-        write_trace_rig(home, "studio-rig", "studio", component_dir.path());
-
-        let rig_path = home
-            .path()
-            .join(".config")
-            .join("homeboy")
-            .join("rigs")
-            .join("studio-rig.json");
-        let mut rig_json: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&rig_path).expect("read rig"))
-                .expect("parse rig");
-        rig_json["trace_guardrails"] = serde_json::json!([
-            { "label": "behavior smoke", "command": "sh -c 'printf regression >&2; exit 7'" }
-        ]);
-        fs::write(
-            &rig_path,
-            serde_json::to_string_pretty(&rig_json).expect("serialize rig"),
-        )
-        .expect("write rig");
-
-        let (output, exit_code) = run_repeat(trace_args_for_rig("studio-rig"))
-            .expect("trace aggregate should preserve artifacts");
-
-        assert_eq!(exit_code, 1);
-        match output {
-            TraceCommandOutput::Aggregate(aggregate) => {
-                assert!(!aggregate.passed);
-                assert_eq!(aggregate.status, "fail");
-                assert_eq!(aggregate.failure_count, 0);
-                assert_eq!(aggregate.guardrail_failure_count, 1);
-                assert_eq!(aggregate.runs.len(), 1);
-                assert_eq!(aggregate.runs[0].status, "pass");
-                assert!(!aggregate.runs[0].artifact_path.is_empty());
-                assert_eq!(aggregate.spans[0].id, "boot_to_ready");
-                assert_eq!(aggregate.spans[0].median_ms, Some(125));
-                assert_eq!(aggregate.guardrails[0].status, "fail");
-                assert!(aggregate.guardrails[0]
-                    .failure
-                    .as_deref()
-                    .unwrap_or_default()
-                    .contains("regression"));
-
-                let artifact = serde_json::to_value(&aggregate).expect("serialize aggregate");
-                assert_eq!(artifact["guardrail_failure_count"], 1);
-                assert_eq!(artifact["guardrails"][0]["label"], "behavior smoke");
-                assert_eq!(artifact["runs"][0]["status"], "pass");
-            }
-            _ => panic!("expected aggregate output"),
         }
     });
 }
